% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiSpMaturityDelayStability-package.R
\docType{package}
\name{MultiSppMaturityDelayStability-package}
\alias{MultiSppMaturityDelayStability}
\alias{MultiSppMaturityDelayStability-package}
\title{Stability Analysis for Multispecies Model with Delayed Maturity}
\description{
Numerical and graphical analyses are provided to reproduce the numerical
demonstration of the local stability analysis for three species of competing
mosquito species with delayed maturity as supplementary material for the
article Stability of difference equations with interspecific density
dependence, competition, and maturation delays.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/csiro-risk-assessment/MultiSppMaturityDelayStability}
}

}
\author{
\strong{Maintainer}: Geoffrey R. Hosack \email{geoff.hosack@csiro.au} (\href{https://orcid.org/0000-0002-6462-6817}{ORCID})

}
\keyword{internal}
